# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Non-Profit Impact Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List all programs
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Create a new program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get program by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProgram
      tags: [programs]
      summary: Update a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "200":
          description: Updated program
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProgram
      tags: [programs]
      summary: Delete a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Funding Entries ────────────────────────────────────────────
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Authenticate and receive a session token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get the authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: string
        fields:
          type: object
          additionalProperties:
            type: string
      required: [error]

    # ── Auth ───────────────────────────────────────────────────────
    LoginInput:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        role:
          type: string
      required: [id, email, role]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/AuthUser"
      required: [token, user]
...
    Program:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
        annualBudget:
          type: number
        peopleServedTarget:
          type: integer
        peopleServedActual:
          type: integer
        outcomesTarget:
          type: integer
        outcomesActual:
          type: integer
        costPerOutcome:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, description, category, status, annualBudget, peopleServedTarget, peopleServedActual, outcomesTarget, outcomesActual]
    ProgramInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
        status:
          type: string
          enum: [active, paused, completed]
        annualBudget:
          type: number
        peopleServedTarget:
          type: integer
        peopleServedActual:
          type: integer
        outcomesTarget:
          type: integer
        outcomesActual:
          type: integer
        costPerOutcome:
          type: number
        startDate:
          type: string
        endDate:
          type: string
      required: [name, description, category, status, annualBudget, peopleServedTarget, peopleServedActual, outcomesTarget, outcomesActual]

    # ── Funding Entry ─────────────────────────────────────────────
...