# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Non-Profit Impact Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  email: string;
  name?: string | null;
  role: string;
}