# lib/api-zod/src/generated/types/programInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Non-Profit Impact Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ProgramInputStatus } from './programInputStatus';

export interface ProgramInput {
  name: string;
  description: string;
  category: string;
  status: ProgramInputStatus;
  annualBudget: number;
  peopleServedTarget: number;
  peopleServedActual: number;
  outcomesTarget: number;
  outcomesActual: number;
  costPerOutcome?: number;
  startDate?: string;
  endDate?: string;
}