# lib/api-zod/src/generated/types/validationErrorResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Non-Profit Impact Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ValidationErrorResponseFields } from './validationErrorResponseFields';

export interface ValidationErrorResponse {
  error: string;
  fields?: ValidationErrorResponseFields;
}